Forest Level Update
Added a bunch of stuff since last push.
Turning everything into a prefab makes the scene file into less than 11mb (not sure if this is a viable option, but it can be unpacked if changes are needed. Unpacking the "forest" prefab makes the scene to around 19mb).

Fog and rain particle system has been added for prototype testing. Changes will definitely be made later on, based on gameplay testing.

New trees with new textures, new grass and dirt textures (placeholders), and updated buildings with interiors. Campfires are placed in different areas, as well as two types of trees in various locations to interact with (red = hitting the tree drops an item, blue = cutting down the tree to block a pathway).