Update 10/28
Imported car, there's now a crash site with lights.
Created a ladder, it is used for getting over holes, like at the look out tower.
Crates are placed in different areas.

Generator parts are placed throughout the whole map now instead of around the lookout tower.

2nd Generator is placed at the lodge, with another laptop and additional lights to turn on (script needs to be changed later on, since this 2nd generator needs 2 parts to turn on, but if the player had already picked up 3 parts to power the generator at the lookout tower, they will also be able to turn on this generator).

Also at the lodge, the front door is blocked off. Players will need to find a way to break through the window (there will be crates).

Added a tent at a campsite with a key to the gate that's blocking off the lake