Lots of stuff
Created a copy of the 11-1 build scene into my folder under "Beta Level".
Adjusted lighting to make it easier for players to see in the dark (still testing).
Adding in more models from Suan.

Added signs at buildings to make it easier for players to know which building is which.
Adjusted scaling for some of the buildings and objects, as they seemed a bit too big/tall for the player.

Adjusted eclipse and flashlight timer, needs more playtesting.
Fixed the issue with the rain passing through buildings.

Adding more lights to some areas.
There's now a smoke particle effect on the campfires whenever you light them.