update for today
Added sticky note interaction, added a code in the generator to destroy a gameobject.

Changed the lookout tower, it now has an electronic lock that can only be open when the generator is powered. There is no lookout tower key, but now there's a lodge key that's in the tower, rewarding the player instead of it just being text on a computer as the goal of the generator