Footsteps are somehow working nearly flawlessly.
Whatever demon has been haunting this project is finally taking pity on me. I changed absolutely nothing since when I left yesterday and yet somehow it just works. I do not understand, but I am not complaining. No rigidbodies are needed on houses, leaving it as it is and saving here.