Fixing some bugs
Made sure that all of the barrels and windows (except in fire lookout tower) are set to be breakable. Setting them to their respectful tags (Barrel and Window) seems to play their specific sounds instead of the "breakable" tag which just doesn't play anything. Need to confirm with Ronan tomorrow.
Fixed the typo on one of the sticky notes in the lookout tower.