Temporary Fix to Generator Issue
I can't seem to replicate the issue on my end, but in Issue SCRUM-276 Ronan was able to get the phone to display 254 gas cans. He didn't note if he tried the debug command for bullets and had an issue there, so I am assuming somewhere along the line there is an negative number showing up somewhere. I've updated the generator script to include a sanity check.